Setting the timezone for sampling date and frozen time for samples to the local timezone of the server.
