+++ baobab/lims/browser/samplebatch/batch.py
@@ -81,12 +81,12 @@ class BatchView(BrowserView):
 
         self.location = ','.join(location_paths)
 
-        self.creation_date = context.getDateCreated().strftime("%Y/%m/%d %H:%M")
+        self.creation_date = context.getDateCreated()    # .strftime("%Y/%m/%d %H:%M")
 
         self.serumColour = context.getField('SerumColour').get(context)
 
         try:
-            self.contrifugation_date = context.getCfgDateTime().strftime("%Y/%m/%d %H:%M")
+            self.contrifugation_date = context.getCfgDateTime()   # .strftime("%Y/%m/%d %H:%M")
         except:
             self.contrifugation_date = ''
